Struct aws_sdk_directory::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for DirectoryService_20150416.

This client allows ergonomic access to a DirectoryService_20150416-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_shared_directory(&self) -> AcceptSharedDirectory<C, M, R>

pub fn add_ip_routes(&self) -> AddIpRoutes<C, M, R>

pub fn add_region(&self) -> AddRegion<C, M, R>

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

pub fn cancel_schema_extension(&self) -> CancelSchemaExtension<C, M, R>

pub fn connect_directory(&self) -> ConnectDirectory<C, M, R>

pub fn create_alias(&self) -> CreateAlias<C, M, R>

pub fn create_computer(&self) -> CreateComputer<C, M, R>

pub fn create_conditional_forwarder(
&self
) -> CreateConditionalForwarder<C, M, R>

pub fn create_directory(&self) -> CreateDirectory<C, M, R>

pub fn create_log_subscription(&self) -> CreateLogSubscription<C, M, R>

pub fn create_microsoft_ad(&self) -> CreateMicrosoftAD<C, M, R>

pub fn create_snapshot(&self) -> CreateSnapshot<C, M, R>

pub fn create_trust(&self) -> CreateTrust<C, M, R>

pub fn delete_conditional_forwarder(
&self
) -> DeleteConditionalForwarder<C, M, R>

pub fn delete_directory(&self) -> DeleteDirectory<C, M, R>

pub fn delete_log_subscription(&self) -> DeleteLogSubscription<C, M, R>

pub fn delete_snapshot(&self) -> DeleteSnapshot<C, M, R>

pub fn delete_trust(&self) -> DeleteTrust<C, M, R>

pub fn deregister_certificate(&self) -> DeregisterCertificate<C, M, R>

pub fn deregister_event_topic(&self) -> DeregisterEventTopic<C, M, R>

pub fn describe_certificate(&self) -> DescribeCertificate<C, M, R>

pub fn describe_client_authentication_settings(
&self
) -> DescribeClientAuthenticationSettings<C, M, R>

pub fn describe_conditional_forwarders(
&self
) -> DescribeConditionalForwarders<C, M, R>

pub fn describe_directories(&self) -> DescribeDirectories<C, M, R>

pub fn describe_domain_controllers(&self) -> DescribeDomainControllers<C, M, R>

pub fn describe_event_topics(&self) -> DescribeEventTopics<C, M, R>

pub fn describe_ldaps_settings(&self) -> DescribeLDAPSSettings<C, M, R>

pub fn describe_regions(&self) -> DescribeRegions<C, M, R>

pub fn describe_shared_directories(&self) -> DescribeSharedDirectories<C, M, R>

pub fn describe_snapshots(&self) -> DescribeSnapshots<C, M, R>

pub fn describe_trusts(&self) -> DescribeTrusts<C, M, R>

pub fn disable_client_authentication(
&self
) -> DisableClientAuthentication<C, M, R>

pub fn disable_ldaps(&self) -> DisableLDAPS<C, M, R>

pub fn disable_radius(&self) -> DisableRadius<C, M, R>

pub fn disable_sso(&self) -> DisableSso<C, M, R>

pub fn enable_client_authentication(
&self
) -> EnableClientAuthentication<C, M, R>

pub fn enable_ldaps(&self) -> EnableLDAPS<C, M, R>

pub fn enable_radius(&self) -> EnableRadius<C, M, R>

pub fn enable_sso(&self) -> EnableSso<C, M, R>

pub fn get_directory_limits(&self) -> GetDirectoryLimits<C, M, R>

pub fn get_snapshot_limits(&self) -> GetSnapshotLimits<C, M, R>

pub fn list_certificates(&self) -> ListCertificates<C, M, R>

pub fn list_ip_routes(&self) -> ListIpRoutes<C, M, R>

pub fn list_log_subscriptions(&self) -> ListLogSubscriptions<C, M, R>

pub fn list_schema_extensions(&self) -> ListSchemaExtensions<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn register_certificate(&self) -> RegisterCertificate<C, M, R>

pub fn register_event_topic(&self) -> RegisterEventTopic<C, M, R>

pub fn reject_shared_directory(&self) -> RejectSharedDirectory<C, M, R>

pub fn remove_ip_routes(&self) -> RemoveIpRoutes<C, M, R>

pub fn remove_region(&self) -> RemoveRegion<C, M, R>

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

pub fn reset_user_password(&self) -> ResetUserPassword<C, M, R>

pub fn restore_from_snapshot(&self) -> RestoreFromSnapshot<C, M, R>

pub fn share_directory(&self) -> ShareDirectory<C, M, R>

pub fn start_schema_extension(&self) -> StartSchemaExtension<C, M, R>

pub fn unshare_directory(&self) -> UnshareDirectory<C, M, R>

pub fn update_conditional_forwarder(
&self
) -> UpdateConditionalForwarder<C, M, R>

pub fn update_number_of_domain_controllers(
&self
) -> UpdateNumberOfDomainControllers<C, M, R>

pub fn update_radius(&self) -> UpdateRadius<C, M, R>

pub fn update_trust(&self) -> UpdateTrust<C, M, R>

pub fn verify_trust(&self) -> VerifyTrust<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self