Module aws_sdk_directory::error

Modules

access_denied_exception

See AccessDeniedException

authentication_failed_exception

See AuthenticationFailedException

certificate_already_exists_exception

See CertificateAlreadyExistsException

certificate_does_not_exist_exception

See CertificateDoesNotExistException

certificate_in_use_exception

See CertificateInUseException

certificate_limit_exceeded_exception

See CertificateLimitExceededException

client_exception

See ClientException

directory_already_in_region_exception

See DirectoryAlreadyInRegionException

directory_already_shared_exception

See DirectoryAlreadySharedException

directory_does_not_exist_exception

See DirectoryDoesNotExistException

directory_limit_exceeded_exception

See DirectoryLimitExceededException

directory_not_shared_exception

See DirectoryNotSharedException

directory_unavailable_exception

See DirectoryUnavailableException

domain_controller_limit_exceeded_exception

See DomainControllerLimitExceededException

entity_already_exists_exception

See EntityAlreadyExistsException

entity_does_not_exist_exception

See EntityDoesNotExistException

insufficient_permissions_exception

See InsufficientPermissionsException

invalid_certificate_exception

See InvalidCertificateException

invalid_client_auth_status_exception

See InvalidClientAuthStatusException

invalid_ldaps_status_exception

See InvalidLdapsStatusException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_password_exception

See InvalidPasswordException

invalid_target_exception

See InvalidTargetException

ip_route_limit_exceeded_exception

See IpRouteLimitExceededException

no_available_certificate_exception

See NoAvailableCertificateException

organizations_exception

See OrganizationsException

region_limit_exceeded_exception

See RegionLimitExceededException

service_exception

See ServiceException

share_limit_exceeded_exception

See ShareLimitExceededException

snapshot_limit_exceeded_exception

See SnapshotLimitExceededException

tag_limit_exceeded_exception

See TagLimitExceededException

unsupported_operation_exception

See UnsupportedOperationException

user_does_not_exist_exception

See UserDoesNotExistException

Structs

AcceptSharedDirectoryError

AccessDeniedException

Client authentication is not available in this region at this time.

AddIpRoutesError

AddRegionError

AddTagsToResourceError

AuthenticationFailedException

An authentication error occurred.

CancelSchemaExtensionError

CertificateAlreadyExistsException

The certificate has already been registered into the system.

CertificateDoesNotExistException

The certificate is not present in the system for describe or deregister activities.

CertificateInUseException

The certificate is being used for the LDAP security connection and cannot be removed

CertificateLimitExceededException

The certificate could not be added because the certificate limit has been reached.

ClientException

A client exception has occurred.

ConnectDirectoryError

CreateAliasError

CreateComputerError

CreateConditionalForwarderError

CreateDirectoryError

CreateLogSubscriptionError

CreateMicrosoftADError

CreateSnapshotError

CreateTrustError

DeleteConditionalForwarderError

DeleteDirectoryError

DeleteLogSubscriptionError

DeleteSnapshotError

DeleteTrustError

DeregisterCertificateError

DeregisterEventTopicError

DescribeCertificateError

DescribeClientAuthenticationSettingsError

DescribeConditionalForwardersError

DescribeDirectoriesError

DescribeDomainControllersError

DescribeEventTopicsError

DescribeLDAPSSettingsError

DescribeRegionsError

DescribeSharedDirectoriesError

DescribeSnapshotsError

DescribeTrustsError

DirectoryAlreadyInRegionException

The Region you specified is the same Region where the Managed Microsoft AD directory

DirectoryAlreadySharedException

The specified directory has already been shared with this Amazon Web Services account.

DirectoryDoesNotExistException

The specified directory does not exist in the system.

DirectoryLimitExceededException

The maximum number of directories in the region has been reached. You can use the

DirectoryNotSharedException

The specified directory has not been shared with this Amazon Web Services account.

DirectoryUnavailableException

The specified directory is unavailable or could not be found.

DisableClientAuthenticationError

DisableLDAPSError

DisableRadiusError

DisableSsoError

DomainControllerLimitExceededException

The maximum allowed number of domain controllers per directory was exceeded. The

EnableClientAuthenticationError

EnableLDAPSError

EnableRadiusError

EnableSsoError

EntityAlreadyExistsException

The specified entity already exists.

EntityDoesNotExistException

The specified entity could not be found.

GetDirectoryLimitsError

GetSnapshotLimitsError

InsufficientPermissionsException

The account does not have sufficient permission to perform the operation.

InvalidCertificateException

The certificate PEM that was provided has incorrect encoding.

InvalidClientAuthStatusException

Client authentication is already enabled.

InvalidLdapsStatusException

The LDAP activities could not be performed because they are limited by the LDAPS

InvalidNextTokenException

The NextToken value is not valid.

InvalidParameterException

One or more parameters are not valid.

InvalidPasswordException

The new password provided by the user does not meet the password complexity

InvalidTargetException

The specified shared target is not valid.

IpRouteLimitExceededException

The maximum allowed number of IP addresses was exceeded. The default limit is 100 IP

ListCertificatesError

ListIpRoutesError

ListLogSubscriptionsError

ListSchemaExtensionsError

ListTagsForResourceError

NoAvailableCertificateException

Client authentication setup could not be completed because at least one valid certificate must be

OrganizationsException

Exception encountered while trying to access your Amazon Web Services organization.

RegionLimitExceededException

You have reached the limit for maximum number of simultaneous Region replications per

RegisterCertificateError

RegisterEventTopicError

RejectSharedDirectoryError

RemoveIpRoutesError

RemoveRegionError

RemoveTagsFromResourceError

ResetUserPasswordError

RestoreFromSnapshotError

ServiceException

An exception has occurred in Directory Service.

ShareDirectoryError

ShareLimitExceededException

The maximum number of Amazon Web Services accounts that you can share with this directory has been

SnapshotLimitExceededException

The maximum number of manual snapshots for the directory has been reached. You can

StartSchemaExtensionError

TagLimitExceededException

The maximum allowed number of tags was exceeded.

UnshareDirectoryError

UnsupportedOperationException

The operation is not supported.

UpdateConditionalForwarderError

UpdateNumberOfDomainControllersError

UpdateRadiusError

UpdateTrustError

UserDoesNotExistException

The user provided a username that does not exist in your directory.

VerifyTrustError

Enums

AcceptSharedDirectoryErrorKind

AddIpRoutesErrorKind

AddRegionErrorKind

AddTagsToResourceErrorKind

CancelSchemaExtensionErrorKind

ConnectDirectoryErrorKind

CreateAliasErrorKind

CreateComputerErrorKind

CreateConditionalForwarderErrorKind

CreateDirectoryErrorKind

CreateLogSubscriptionErrorKind

CreateMicrosoftADErrorKind

CreateSnapshotErrorKind

CreateTrustErrorKind

DeleteConditionalForwarderErrorKind

DeleteDirectoryErrorKind

DeleteLogSubscriptionErrorKind

DeleteSnapshotErrorKind

DeleteTrustErrorKind

DeregisterCertificateErrorKind

DeregisterEventTopicErrorKind

DescribeCertificateErrorKind

DescribeClientAuthenticationSettingsErrorKind

DescribeConditionalForwardersErrorKind

DescribeDirectoriesErrorKind

DescribeDomainControllersErrorKind

DescribeEventTopicsErrorKind

DescribeLDAPSSettingsErrorKind

DescribeRegionsErrorKind

DescribeSharedDirectoriesErrorKind

DescribeSnapshotsErrorKind

DescribeTrustsErrorKind

DisableClientAuthenticationErrorKind

DisableLDAPSErrorKind

DisableRadiusErrorKind

DisableSsoErrorKind

EnableClientAuthenticationErrorKind

EnableLDAPSErrorKind

EnableRadiusErrorKind

EnableSsoErrorKind

GetDirectoryLimitsErrorKind

GetSnapshotLimitsErrorKind

ListCertificatesErrorKind

ListIpRoutesErrorKind

ListLogSubscriptionsErrorKind

ListSchemaExtensionsErrorKind

ListTagsForResourceErrorKind

RegisterCertificateErrorKind

RegisterEventTopicErrorKind

RejectSharedDirectoryErrorKind

RemoveIpRoutesErrorKind

RemoveRegionErrorKind

RemoveTagsFromResourceErrorKind

ResetUserPasswordErrorKind

RestoreFromSnapshotErrorKind

ShareDirectoryErrorKind

StartSchemaExtensionErrorKind

UnshareDirectoryErrorKind

UpdateConditionalForwarderErrorKind

UpdateNumberOfDomainControllersErrorKind

UpdateRadiusErrorKind

UpdateTrustErrorKind

VerifyTrustErrorKind