Module aws_sdk_directory::input

Modules

accept_shared_directory_input

See AcceptSharedDirectoryInput

add_ip_routes_input

See AddIpRoutesInput

add_region_input

See AddRegionInput

add_tags_to_resource_input

See AddTagsToResourceInput

cancel_schema_extension_input

See CancelSchemaExtensionInput

connect_directory_input

See ConnectDirectoryInput

create_alias_input

See CreateAliasInput

create_computer_input

See CreateComputerInput

create_conditional_forwarder_input

See CreateConditionalForwarderInput

create_directory_input

See CreateDirectoryInput

create_log_subscription_input

See CreateLogSubscriptionInput

create_microsoft_ad_input

See CreateMicrosoftAdInput

create_snapshot_input

See CreateSnapshotInput

create_trust_input

See CreateTrustInput

delete_conditional_forwarder_input

See DeleteConditionalForwarderInput

delete_directory_input

See DeleteDirectoryInput

delete_log_subscription_input

See DeleteLogSubscriptionInput

delete_snapshot_input

See DeleteSnapshotInput

delete_trust_input

See DeleteTrustInput

deregister_certificate_input

See DeregisterCertificateInput

deregister_event_topic_input

See DeregisterEventTopicInput

describe_certificate_input

See DescribeCertificateInput

describe_client_authentication_settings_input

See DescribeClientAuthenticationSettingsInput

describe_conditional_forwarders_input

See DescribeConditionalForwardersInput

describe_directories_input

See DescribeDirectoriesInput

describe_domain_controllers_input

See DescribeDomainControllersInput

describe_event_topics_input

See DescribeEventTopicsInput

describe_ldaps_settings_input

See DescribeLdapsSettingsInput

describe_regions_input

See DescribeRegionsInput

describe_shared_directories_input

See DescribeSharedDirectoriesInput

describe_snapshots_input

See DescribeSnapshotsInput

describe_trusts_input

See DescribeTrustsInput

disable_client_authentication_input

See DisableClientAuthenticationInput

disable_ldaps_input

See DisableLdapsInput

disable_radius_input

See DisableRadiusInput

disable_sso_input

See DisableSsoInput

enable_client_authentication_input

See EnableClientAuthenticationInput

enable_ldaps_input

See EnableLdapsInput

enable_radius_input

See EnableRadiusInput

enable_sso_input

See EnableSsoInput

get_directory_limits_input

See GetDirectoryLimitsInput

get_snapshot_limits_input

See GetSnapshotLimitsInput

list_certificates_input

See ListCertificatesInput

list_ip_routes_input

See ListIpRoutesInput

list_log_subscriptions_input

See ListLogSubscriptionsInput

list_schema_extensions_input

See ListSchemaExtensionsInput

list_tags_for_resource_input

See ListTagsForResourceInput

register_certificate_input

See RegisterCertificateInput

register_event_topic_input

See RegisterEventTopicInput

reject_shared_directory_input

See RejectSharedDirectoryInput

remove_ip_routes_input

See RemoveIpRoutesInput

remove_region_input

See RemoveRegionInput

remove_tags_from_resource_input

See RemoveTagsFromResourceInput

reset_user_password_input

See ResetUserPasswordInput

restore_from_snapshot_input

See RestoreFromSnapshotInput

share_directory_input

See ShareDirectoryInput

start_schema_extension_input

See StartSchemaExtensionInput

unshare_directory_input

See UnshareDirectoryInput

update_conditional_forwarder_input

See UpdateConditionalForwarderInput

update_number_of_domain_controllers_input

See UpdateNumberOfDomainControllersInput

update_radius_input

See UpdateRadiusInput

update_trust_input

See UpdateTrustInput

verify_trust_input

See VerifyTrustInput

Structs

AcceptSharedDirectoryInput

AddIpRoutesInput

AddRegionInput

AddTagsToResourceInput

CancelSchemaExtensionInput

ConnectDirectoryInput

Contains the inputs for the ConnectDirectory operation.

CreateAliasInput

Contains the inputs for the CreateAlias operation.

CreateComputerInput

Contains the inputs for the CreateComputer operation.

CreateConditionalForwarderInput

Initiates the creation of a conditional forwarder for your Directory Service for Microsoft Active

CreateDirectoryInput

Contains the inputs for the CreateDirectory operation.

CreateLogSubscriptionInput

CreateMicrosoftAdInput

Creates an Managed Microsoft AD directory.

CreateSnapshotInput

Contains the inputs for the CreateSnapshot operation.

CreateTrustInput

Directory Service for Microsoft Active Directory allows you to configure trust relationships. For

DeleteConditionalForwarderInput

Deletes a conditional forwarder.

DeleteDirectoryInput

Contains the inputs for the DeleteDirectory operation.

DeleteLogSubscriptionInput

DeleteSnapshotInput

Contains the inputs for the DeleteSnapshot operation.

DeleteTrustInput

Deletes the local side of an existing trust relationship between the Managed Microsoft AD

DeregisterCertificateInput

DeregisterEventTopicInput

Removes the specified directory as a publisher to the specified Amazon SNS topic.

DescribeCertificateInput

DescribeClientAuthenticationSettingsInput

DescribeConditionalForwardersInput

Describes a conditional forwarder.

DescribeDirectoriesInput

Contains the inputs for the DescribeDirectories operation.

DescribeDomainControllersInput

DescribeEventTopicsInput

Describes event topics.

DescribeLdapsSettingsInput

DescribeRegionsInput

DescribeSharedDirectoriesInput

DescribeSnapshotsInput

Contains the inputs for the DescribeSnapshots operation.

DescribeTrustsInput

Describes the trust relationships for a particular Managed Microsoft AD directory. If no input

DisableClientAuthenticationInput

DisableLdapsInput

DisableRadiusInput

Contains the inputs for the DisableRadius operation.

DisableSsoInput

Contains the inputs for the DisableSso operation.

EnableClientAuthenticationInput

EnableLdapsInput

EnableRadiusInput

Contains the inputs for the EnableRadius operation.

EnableSsoInput

Contains the inputs for the EnableSso operation.

GetDirectoryLimitsInput

Contains the inputs for the GetDirectoryLimits operation.

GetSnapshotLimitsInput

Contains the inputs for the GetSnapshotLimits operation.

ListCertificatesInput

ListIpRoutesInput

ListLogSubscriptionsInput

ListSchemaExtensionsInput

ListTagsForResourceInput

RegisterCertificateInput

RegisterEventTopicInput

Registers a new event topic.

RejectSharedDirectoryInput

RemoveIpRoutesInput

RemoveRegionInput

RemoveTagsFromResourceInput

ResetUserPasswordInput

RestoreFromSnapshotInput

An object representing the inputs for the RestoreFromSnapshot operation.

ShareDirectoryInput

StartSchemaExtensionInput

UnshareDirectoryInput

UpdateConditionalForwarderInput

Updates a conditional forwarder.

UpdateNumberOfDomainControllersInput

UpdateRadiusInput

Contains the inputs for the UpdateRadius operation.

UpdateTrustInput

VerifyTrustInput

Initiates the verification of an existing trust relationship between an Managed Microsoft AD