Module aws_sdk_directory::model

Modules

attribute

See Attribute

certificate

See Certificate

certificate_info

See CertificateInfo

client_authentication_setting_info

See ClientAuthenticationSettingInfo

client_cert_auth_settings

See ClientCertAuthSettings

computer

See Computer

conditional_forwarder

See ConditionalForwarder

directory_connect_settings

See DirectoryConnectSettings

directory_connect_settings_description

See DirectoryConnectSettingsDescription

directory_description

See DirectoryDescription

directory_limits

See DirectoryLimits

directory_vpc_settings

See DirectoryVpcSettings

directory_vpc_settings_description

See DirectoryVpcSettingsDescription

domain_controller

See DomainController

event_topic

See EventTopic

ip_route

See IpRoute

ip_route_info

See IpRouteInfo

ldaps_setting_info

See LdapsSettingInfo

log_subscription

See LogSubscription

owner_directory_description

See OwnerDirectoryDescription

radius_settings

See RadiusSettings

region_description

See RegionDescription

regions_info

See RegionsInfo

schema_extension_info

See SchemaExtensionInfo

share_target

See ShareTarget

shared_directory

See SharedDirectory

snapshot

See Snapshot

snapshot_limits

See SnapshotLimits

tag

See Tag

trust

See Trust

unshare_target

See UnshareTarget

Structs

Attribute

Represents a named directory attribute.

Certificate

Information about the certificate.

CertificateInfo

Contains general information about a certificate.

ClientAuthenticationSettingInfo

Contains information about a client authentication method for a directory.

ClientCertAuthSettings

Contains information about the client certificate authentication settings for the RegisterCertificate and DescribeCertificate operations.

Computer

Contains information about a computer account in a directory.

ConditionalForwarder

Points to a remote domain with which you are setting up a trust relationship.

DirectoryConnectSettings

Contains information for the ConnectDirectory operation when an AD

DirectoryConnectSettingsDescription

Contains information about an AD Connector directory.

DirectoryDescription

Contains information about an Directory Service directory.

DirectoryLimits

Contains directory limit information for a Region.

DirectoryVpcSettings

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

DirectoryVpcSettingsDescription

Contains information about the directory.

DomainController

Contains information about the domain controllers for a specified directory.

EventTopic

Information about Amazon SNS topic and Directory Service directory associations.

IpRoute

IP address block. This is often the address block of the DNS server used for your

IpRouteInfo

Information about one or more IP address blocks.

LdapsSettingInfo

Contains general information about the LDAPS settings.

LogSubscription

Represents a log subscription, which tracks real-time data from a chosen log group to a

OwnerDirectoryDescription

Describes the directory owner account details that have been shared to the directory

RadiusSettings

Contains information about a Remote Authentication Dial In User Service (RADIUS)

RegionDescription

The replicated Region information for a directory.

RegionsInfo

Provides information about the Regions that are configured for multi-Region

SchemaExtensionInfo

Information about a schema extension.

ShareTarget

Identifier that contains details about the directory consumer account.

SharedDirectory

Details about the shared directory in the directory owner account for which the share

Snapshot

Describes a directory snapshot.

SnapshotLimits

Contains manual snapshot limit information for a directory.

Tag

Metadata assigned to a directory consisting of a key-value pair.

Trust

Describes a trust relationship between an Managed Microsoft AD directory and an external

UnshareTarget

Identifier that contains details about the directory consumer account with whom the

Enums

CertificateState

CertificateType

ClientAuthenticationStatus

ClientAuthenticationType

DirectoryEdition

DirectorySize

DirectoryStage

DirectoryType

DomainControllerStatus

IpRouteStatusMsg

LdapsStatus

LdapsType

RadiusAuthenticationProtocol

RadiusStatus

RegionType

ReplicationScope

SchemaExtensionStatus

SelectiveAuth

ShareMethod

ShareStatus

SnapshotStatus

SnapshotType

TargetType

TopicStatus

TrustDirection

TrustState

TrustType