Struct aws_sdk_directory::model::directory_description::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DirectoryDescription

Implementations

impl Builder

pub fn directory_id(self, input: impl Into<String>) -> Self

The directory identifier.

pub fn set_directory_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The fully qualified name of the directory.

pub fn set_name(self, input: Option<String>) -> Self

pub fn short_name(self, input: impl Into<String>) -> Self

The short name of the directory.

pub fn set_short_name(self, input: Option<String>) -> Self

pub fn size(self, input: DirectorySize) -> Self

The directory size.

pub fn set_size(self, input: Option<DirectorySize>) -> Self

pub fn edition(self, input: DirectoryEdition) -> Self

The edition associated with this directory.

pub fn set_edition(self, input: Option<DirectoryEdition>) -> Self

pub fn alias(self, input: impl Into<String>) -> Self

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

pub fn set_alias(self, input: Option<String>) -> Self

pub fn access_url(self, input: impl Into<String>) -> Self

The access URL for the directory, such as http://.awsapps.com. If no alias has been created for the directory, is the directory identifier, such as d-XXXXXXXXXX.

pub fn set_access_url(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description for the directory.

pub fn set_description(self, input: Option<String>) -> Self

pub fn dns_ip_addrs(self, input: impl Into<String>) -> Self

pub fn set_dns_ip_addrs(self, input: Option<Vec<String>>) -> Self

pub fn stage(self, input: DirectoryStage) -> Self

The current stage of the directory.

pub fn set_stage(self, input: Option<DirectoryStage>) -> Self

pub fn share_status(self, input: ShareStatus) -> Self

Current directory status of the shared Managed Microsoft AD directory.

pub fn set_share_status(self, input: Option<ShareStatus>) -> Self

pub fn share_method(self, input: ShareMethod) -> Self

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

pub fn set_share_method(self, input: Option<ShareMethod>) -> Self

pub fn share_notes(self, input: impl Into<String>) -> Self

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

pub fn set_share_notes(self, input: Option<String>) -> Self

pub fn launch_time(self, input: Instant) -> Self

Specifies when the directory was created.

pub fn set_launch_time(self, input: Option<Instant>) -> Self

pub fn stage_last_updated_date_time(self, input: Instant) -> Self

The date and time that the stage was last updated.

pub fn set_stage_last_updated_date_time(self, input: Option<Instant>) -> Self

pub fn type(self, input: DirectoryType) -> Self

The directory size.

pub fn set_type(self, input: Option<DirectoryType>) -> Self

pub fn vpc_settings(self, input: DirectoryVpcSettingsDescription) -> Self

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

pub fn set_vpc_settings(
self,
input: Option<DirectoryVpcSettingsDescription>
) -> Self

pub fn connect_settings(
self,
input: DirectoryConnectSettingsDescription
) -> Self

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

pub fn set_connect_settings(
self,
input: Option<DirectoryConnectSettingsDescription>
) -> Self

pub fn radius_settings(self, input: RadiusSettings) -> Self

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

pub fn set_radius_settings(self, input: Option<RadiusSettings>) -> Self

pub fn radius_status(self, input: RadiusStatus) -> Self

The status of the RADIUS MFA server connection.

pub fn set_radius_status(self, input: Option<RadiusStatus>) -> Self

pub fn stage_reason(self, input: impl Into<String>) -> Self

Additional information about the directory stage.

pub fn set_stage_reason(self, input: Option<String>) -> Self

pub fn sso_enabled(self, input: bool) -> Self

Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.

pub fn set_sso_enabled(self, input: Option<bool>) -> Self

pub fn desired_number_of_domain_controllers(self, input: i32) -> Self

The desired number of domain controllers in the directory if the directory is Microsoft AD.

pub fn set_desired_number_of_domain_controllers(
self,
input: Option<i32>
) -> Self

pub fn owner_directory_description(
self,
input: OwnerDirectoryDescription
) -> Self

Describes the Managed Microsoft AD directory in the directory owner account.

pub fn set_owner_directory_description(
self,
input: Option<OwnerDirectoryDescription>
) -> Self

pub fn regions_info(self, input: RegionsInfo) -> Self

Lists the Regions where the directory has replicated.

pub fn set_regions_info(self, input: Option<RegionsInfo>) -> Self

pub fn build(self) -> DirectoryDescription

Consumes the builder and constructs a DirectoryDescription