Struct aws_sdk_directory::model::DirectoryDescription

#[non_exhaustive]
pub struct DirectoryDescription {

    pub directory_id: Option<String>,
    pub name: Option<String>,
    pub short_name: Option<String>,
    pub size: Option<DirectorySize>,
    pub edition: Option<DirectoryEdition>,
    pub alias: Option<String>,
    pub access_url: Option<String>,
    pub description: Option<String>,
    pub dns_ip_addrs: Option<Vec<String>>,
    pub stage: Option<DirectoryStage>,
    pub share_status: Option<ShareStatus>,
    pub share_method: Option<ShareMethod>,
    pub share_notes: Option<String>,
    pub launch_time: Option<Instant>,
    pub stage_last_updated_date_time: Option<Instant>,
    pub type: Option<DirectoryType>,
    pub vpc_settings: Option<DirectoryVpcSettingsDescription>,
    pub connect_settings: Option<DirectoryConnectSettingsDescription>,
    pub radius_settings: Option<RadiusSettings>,
    pub radius_status: Option<RadiusStatus>,
    pub stage_reason: Option<String>,
    pub sso_enabled: bool,
    pub desired_number_of_domain_controllers: i32,
    pub owner_directory_description: Option<OwnerDirectoryDescription>,
    pub regions_info: Option<RegionsInfo>,

}

Contains information about an Directory Service directory.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

directory_id: Option<String>

The directory identifier.

name: Option<String>

The fully qualified name of the directory.

short_name: Option<String>

The short name of the directory.

size: Option<DirectorySize>

The directory size.

edition: Option<DirectoryEdition>

The edition associated with this directory.

alias: Option<String>

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

access_url: Option<String>

The access URL for the directory, such as http://.awsapps.com. If no alias has been created for the directory, is the directory identifier, such as d-XXXXXXXXXX.

description: Option<String>

The description for the directory.

dns_ip_addrs: Option<Vec<String>>

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in your self-managed directory to which the AD Connector is connected.

stage: Option<DirectoryStage>

The current stage of the directory.

share_status: Option<ShareStatus>

Current directory status of the shared Managed Microsoft AD directory.

share_method: Option<ShareMethod>

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

share_notes: Option<String>

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

launch_time: Option<Instant>

Specifies when the directory was created.

stage_last_updated_date_time: Option<Instant>

The date and time that the stage was last updated.

type: Option<DirectoryType>

The directory size.

vpc_settings: Option<DirectoryVpcSettingsDescription>

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

connect_settings: Option<DirectoryConnectSettingsDescription>

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

radius_settings: Option<RadiusSettings>

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

radius_status: Option<RadiusStatus>

The status of the RADIUS MFA server connection.

stage_reason: Option<String>

Additional information about the directory stage.

sso_enabled: bool

Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.

desired_number_of_domain_controllers: i32

The desired number of domain controllers in the directory if the directory is Microsoft AD.

owner_directory_description: Option<OwnerDirectoryDescription>

Describes the Managed Microsoft AD directory in the directory owner account.

regions_info: Option<RegionsInfo>

Lists the Regions where the directory has replicated.

Implementations

impl DirectoryDescription

pub fn directory_id(&self) -> Option<&str>

The directory identifier.

pub fn name(&self) -> Option<&str>

The fully qualified name of the directory.

pub fn short_name(&self) -> Option<&str>

The short name of the directory.

pub fn size(&self) -> Option<&DirectorySize>

The directory size.

pub fn edition(&self) -> Option<&DirectoryEdition>

The edition associated with this directory.

pub fn alias(&self) -> Option<&str>

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

pub fn access_url(&self) -> Option<&str>

The access URL for the directory, such as http://.awsapps.com. If no alias has been created for the directory, is the directory identifier, such as d-XXXXXXXXXX.

pub fn description(&self) -> Option<&str>

The description for the directory.

pub fn dns_ip_addrs(&self) -> Option<&[String]>

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in your self-managed directory to which the AD Connector is connected.

pub fn stage(&self) -> Option<&DirectoryStage>

The current stage of the directory.

pub fn share_status(&self) -> Option<&ShareStatus>

Current directory status of the shared Managed Microsoft AD directory.

pub fn share_method(&self) -> Option<&ShareMethod>

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

pub fn share_notes(&self) -> Option<&str>

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

pub fn launch_time(&self) -> Option<&Instant>

Specifies when the directory was created.

pub fn stage_last_updated_date_time(&self) -> Option<&Instant>

The date and time that the stage was last updated.

pub fn type(&self) -> Option<&DirectoryType>

The directory size.

pub fn vpc_settings(&self) -> Option<&DirectoryVpcSettingsDescription>

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

pub fn connect_settings(&self) -> Option<&DirectoryConnectSettingsDescription>

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

pub fn radius_settings(&self) -> Option<&RadiusSettings>

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

pub fn radius_status(&self) -> Option<&RadiusStatus>

The status of the RADIUS MFA server connection.

pub fn stage_reason(&self) -> Option<&str>

Additional information about the directory stage.

pub fn sso_enabled(&self) -> bool

Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.

pub fn desired_number_of_domain_controllers(&self) -> i32

The desired number of domain controllers in the directory if the directory is Microsoft AD.

pub fn owner_directory_description(&self) -> Option<&OwnerDirectoryDescription>

Describes the Managed Microsoft AD directory in the directory owner account.

pub fn regions_info(&self) -> Option<&RegionsInfo>

Lists the Regions where the directory has replicated.

impl DirectoryDescription

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DirectoryDescription