Module aws_sdk_directory::operation

Structs

AcceptSharedDirectory

Accepts a directory sharing request that was sent from the directory owner account.

AddIpRoutes

If the DNS server for your self-managed domain uses a publicly addressable IP address,

AddRegion

Adds two domain controllers in the specified Region for the specified directory.

AddTagsToResource

Adds or overwrites one or more tags for the specified directory. Each directory can

CancelSchemaExtension

Cancels an in-progress schema extension to a Microsoft AD directory. Once a schema

ConnectDirectory

Creates an AD Connector to connect to a self-managed directory.

CreateAlias

Creates an alias for a directory and assigns the alias to the directory. The alias is used

CreateComputer

Creates an Active Directory computer object in the specified directory.

CreateConditionalForwarder

Creates a conditional forwarder associated with your Amazon Web Services directory. Conditional

CreateDirectory

Creates a Simple AD directory. For more information, see Simple Active Directory in the Directory Service Admin

CreateLogSubscription

Creates a subscription to forward real-time Directory Service domain controller security

CreateMicrosoftAD

Creates a Microsoft AD directory in the Amazon Web Services Cloud. For more information, see Managed Microsoft AD in the Directory Service Admin Guide.

CreateSnapshot

Creates a snapshot of a Simple AD or Microsoft AD directory in the Amazon Web Services cloud.

CreateTrust

Directory Service for Microsoft Active Directory allows you to configure trust relationships. For

DeleteConditionalForwarder

Deletes a conditional forwarder that has been set up for your Amazon Web Services

DeleteDirectory

Deletes an Directory Service directory.

DeleteLogSubscription

Deletes the specified log subscription.

DeleteSnapshot

Deletes a directory snapshot.

DeleteTrust

Deletes an existing trust relationship between your Managed Microsoft AD directory and an external

DeregisterCertificate

Deletes from the system the certificate that was registered for secure LDAP or client certificate authentication.

DeregisterEventTopic

Removes the specified directory as a publisher to the specified Amazon SNS topic.

DescribeCertificate

Displays information about the certificate registered for secure LDAP or client certificate authentication.

DescribeClientAuthenticationSettings

Retrieves information about the type of client authentication for the specified directory, if the type is specified. If no type is specified, information about all client authentication types that are supported for the specified directory is retrieved. Currently, only SmartCard is supported.

DescribeConditionalForwarders

Obtains information about the conditional forwarders for this account.

DescribeDirectories

Obtains information about the directories that belong to this account.

DescribeDomainControllers

Provides information about any domain controllers in your directory.

DescribeEventTopics

Obtains information about which Amazon SNS topics receive status messages from the specified

DescribeLDAPSSettings

Describes the status of LDAP security for the specified directory.

DescribeRegions

Provides information about the Regions that are configured for multi-Region

DescribeSharedDirectories

Returns the shared directories in your account.

DescribeSnapshots

Obtains information about the directory snapshots that belong to this account.

DescribeTrusts

Obtains information about the trust relationships for this account.

DisableClientAuthentication

Disables alternative client authentication methods for the specified directory.

DisableLDAPS

Deactivates LDAP secure calls for the specified directory.

DisableRadius

Disables multi-factor authentication (MFA) with the Remote Authentication Dial In

DisableSso

Disables single-sign on for a directory.

EnableClientAuthentication

Enables alternative client authentication methods for the specified directory.

EnableLDAPS

Activates the switch for the specific directory to always use LDAP secure calls.

EnableRadius

Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User

EnableSso

Enables single sign-on for a directory. Single sign-on allows users in your directory to

GetDirectoryLimits

Obtains directory limit information for the current Region.

GetSnapshotLimits

Obtains the manual snapshot limits for a directory.

ListCertificates

For the specified directory, lists all the certificates registered for a secure LDAP or client certificate authentication.

ListIpRoutes

Lists the address blocks that you have added to a directory.

ListLogSubscriptions

Lists the active log subscriptions for the Amazon Web Services account.

ListSchemaExtensions

Lists all schema extensions applied to a Microsoft AD Directory.

ListTagsForResource

Lists all tags on a directory.

RegisterCertificate

Registers a certificate for a secure LDAP or client certificate authentication.

RegisterEventTopic

Associates a directory with an Amazon SNS topic. This establishes the directory as a

RejectSharedDirectory

Rejects a directory sharing request that was sent from the directory owner account.

RemoveIpRoutes

Removes IP address blocks from a directory.

RemoveRegion

Stops all replication and removes the domain controllers from the specified Region. You

RemoveTagsFromResource

Removes tags from a directory.

ResetUserPassword

Resets the password for any user in your Managed Microsoft AD or Simple AD

RestoreFromSnapshot

Restores a directory using an existing directory snapshot.

ShareDirectory

Shares a specified directory (DirectoryId) in your Amazon Web Services account (directory

StartSchemaExtension

Applies a schema extension to a Microsoft AD directory.

UnshareDirectory

Stops the directory sharing between the directory owner and consumer accounts.

UpdateConditionalForwarder

Updates a conditional forwarder that has been set up for your Amazon Web Services

UpdateNumberOfDomainControllers

Adds or removes domain controllers to or from the directory. Based on the difference

UpdateRadius

Updates the Remote Authentication Dial In User Service (RADIUS) server information

UpdateTrust

Updates the trust that has been set up between your Managed Microsoft AD directory and an

VerifyTrust

Directory Service for Microsoft Active Directory allows you to configure and verify trust